+++ packages/odyssey-react-mui/src/CheckboxGroup.tsx
@@ -16,7 +16,7 @@ import { memo, ReactElement, useCallback } from "react";
 import { Checkbox } from "./Checkbox";
 import { Field } from "./Field";
 import { FieldComponentProps } from "./FieldComponentProps";
-import type { AllowedProps } from "./AllowedProps";
+import type { HtmlProps } from "./HtmlProps";
 
 export type CheckboxGroupProps = {
   /**
@@ -35,20 +35,24 @@ export type CheckboxGroupProps = {
   label: string;
 } & Pick<
   FieldComponentProps,
+  | "ariaDescribedBy"
   | "errorMessage"
   | "errorMessageList"
   | "hint"
   | "HintLinkComponent"
+  | "id"
   | "isDisabled"
 > &
-  AllowedProps;
+  HtmlProps;
 
 const CheckboxGroup = ({
+  ariaDescribedBy,
   children,
   errorMessage,
   errorMessageList,
   hint,
   HintLinkComponent,
+  id: idOverride,
   isDisabled,
   isRequired = false,
   label,
@@ -56,12 +60,13 @@ const CheckboxGroup = ({
   translate,
 }: CheckboxGroupProps) => {
   const renderFieldComponent = useCallback(
-    ({ ariaDescribedBy, errorMessageElementId, labelElementId }) => (
+    ({ ariaDescribedBy, errorMessageElementId, id, labelElementId }) => (
       <MuiFormGroup
         aria-describedby={ariaDescribedBy}
         aria-errormessage={errorMessageElementId}
         aria-labelledby={labelElementId}
         data-se={testId}
+        id={id}
         translate={translate}
       >
         {children}
@@ -72,12 +77,14 @@ const CheckboxGroup = ({
 
   return (
     <Field
+      ariaDescribedBy={ariaDescribedBy}
       errorMessage={errorMessage}
       errorMessageList={errorMessageList}
       fieldType="group"
       hasVisibleLabel={true}
       hint={hint}
       HintLinkComponent={HintLinkComponent}
+      id={idOverride}
       isDisabled={isDisabled}
       isOptional={!isRequired}
       label={label}