+++ packages/odyssey-react-mui/src/CircularProgress.tsx
@@ -10,9 +10,10 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
+import { memo } from "react";
 import { CircularProgress as MuiCircularProgress } from "@mui/material";
 
-import type { AllowedProps } from "./AllowedProps";
+import type { HtmlProps } from "./HtmlProps";
 
 export type CircularProgressProps = {
   /**
@@ -24,9 +25,9 @@ export type CircularProgressProps = {
    * If undefined, the spinner will spin perpetually.
    */
   value?: number;
-} & AllowedProps;
+} & HtmlProps;
 
-export const CircularProgress = ({
+const CircularProgress = ({
   ariaLabel,
   testId,
   value,
@@ -38,3 +39,8 @@ export const CircularProgress = ({
     aria-label={ariaLabel}
   />
 );
+
+const MemoizedCircularProgress = memo(CircularProgress);
+MemoizedCircularProgress.displayName = "CircularProgress";
+
+export { MemoizedCircularProgress as CircularProgress };